curse_timp/requirements.txt
pycryptodome==3.19.1
tkinter 